Update Entity.php
I'm not sure if it's the best way to return the name of the class in this way, but anyway it will resolve the issue with returning the integer instead of string. Its not critical and it can be reproduced only in case you decide to change the structure of the project.